Underp initialization of system tray module
The module has two dasbus services:
1. The watcher receives registration requests from applications and
   forwards them to the host.
2. The host instantiates and adds the icon to the GUI.

Previously, they start in dedicated threads along with dasbus's main
loops, which is problematic:
1. dasbus's main loop simply wraps GLib.MainLoop, but without creating a
   new GLib.MainContext. As a result, the global default “main” main
   loop is started for three times, which is clearly undesired.
2. The host manipulates Gtk widgets, which should be done in the main
   thread.

To address these issues, this commit directly starts the host in the
main thread and removes its dasbus loop. While the same thing should be
done for the watcher, a dbus freeze is observed when working with
playerctl (see code comment for details). Starting the watcher in
another process effectively avoids the freeze.